+++ backend/pipelined/handle_test.go
@@ -95,8 +95,10 @@ func TestPipelinedHandleEvent(t *testing.T) {
 		return m, nil
 	}
 
-	assert.NoError(t, p.handleEvent(event))
-	m.AssertCalled(t, "HandleEvent", event, mock.Anything)
+	// TODO: Re-enable the following functionality
+	// Feature temporarily disabled: https://github.com/sensu/sensu-go/issues/1883
+	// assert.NoError(t, p.handleEvent(event))
+	// m.AssertCalled(t, "HandleEvent", event, mock.Anything)
 }
 
 func TestPipelinedExpandHandlers(t *testing.T) {